requirements.txt
absl-py==2.0.0
altair==5.1.2
annotated-types==0.6.0
ansi2html==1.8.0
anyio==3.7.1
astunparse==1.6.3
attrs==23.1.0
blinker==1.7.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
exceptiongroup==1.1.3
fastapi==0.104.0
fiona==1.9.5
flask==3.0.0
flatbuffers==23.5.26
fonttools==4.43.1
gast==0.4.0
geopandas==0.14.0
gitdb==4.0.11
GitPython==3.1.40
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==2.28.2
google-auth-oauthlib==1.0.0
google-generativeai==0.4.0
google-pasta==0.2.0
googleapis-common-protos==1.62.0
greenlet==3.0.0
grpcio==1.59.0
h11==0.14.0
h5py==3.10.0
httptools==0.6.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
itsdangerous==2.1.2
jax==0.4.19
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
keras==2.14.0
kiwisolver==1.4.5
libclang==16.0.6
loguru==0.7.2
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
mdurl==0.1.2
ml-dtypes==0.2.0
nest-asyncio==1.5.8
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.1
Pillow==10.1.0
proto-plus==1.23.0
protobuf==4.24.4
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.4.2
pydantic-core==2.10.1
pydeck==0.8.1b0
Pygments==2.16.1
pyparsing==3.1.1
pyproj==3.6.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
rich==13.6.0
rpds-py==0.10.6
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.3
shapely==2.0.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
starlette==0.27.0
streamlit==1.28.0
tenacity==8.2.3
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-intel==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.2
typing-extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.7
uvicorn==0.23.2
validators==0.22.0
watchdog==3.0.0
watchfiles==0.21.0
websockets==12.0
werkzeug==3.0.0
win32-setctime==1.1.0
wrapt==1.14.1
zipp==3.17.0
